+++ auto_route/example/lib/mobile/main.dart
@@ -23,19 +23,16 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return MaterialApp.router(
       theme: ThemeData.dark().copyWith(
-        pageTransitionsTheme: PageTransitionsTheme(
-            builders: {
-              TargetPlatform.iOS: NoShadowCupertinoPageTransitionsBuilder(),
-              TargetPlatform.android: FadeUpwardsPageTransitionsBuilder(),
-            }
-        )
+        pageTransitionsTheme: PageTransitionsTheme(builders: {
+          TargetPlatform.iOS: NoShadowCupertinoPageTransitionsBuilder(),
+          TargetPlatform.android: FadeUpwardsPageTransitionsBuilder(),
+        }),
       ),
       routerDelegate: _rootRouter.delegate(
-        initialDeepLink: '/profile/my-books',
         navigatorObservers: () => [AutoRouteObserver()],
       ),
       routeInformationProvider: _rootRouter.routeInfoProvider(),
-      routeInformationParser: _rootRouter.defaultRouteParser(),
+      routeInformationParser: _rootRouter.defaultRouteParser(includePrefixMatches: true),
       builder: (_, router) {
         return ChangeNotifierProvider<AuthService>(
           create: (_) => authService,