+++ auto_route/example/lib/mobile/router/router.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:example/mobile/screens/books/book_details_page.dart';
 import 'package:example/mobile/screens/books/book_list_page.dart';
+import 'package:flutter/material.dart';
 import '../screens/home_page.dart';
 import '../screens/login_page.dart';
 import '../screens/profile/routes.dart';
@@ -18,7 +19,7 @@ import '../screens/user-data/routes.dart';
       children: [
         AutoRoute(
           path: 'books',
-          page: EmptyRouterPage,
+          page: EmptyRouterScreen,
           name: 'BooksTab',
           initial: true,
           maintainState: true,
@@ -27,12 +28,9 @@ import '../screens/user-data/routes.dart';
               path: '',
               page: BookListScreen,
             ),
-            AutoRoute(
-              path: ':id',
-              page: BookDetailsPage,
-              fullscreenDialog: true,
-              // meta: {'hideBottomNav': true},
-            ),
+            AutoRoute(path: ':id', page: BookDetailsPage, fullscreenDialog: true, children: [AutoRoute(page: InheritedParamScreen)]
+                // meta: {'hideBottomNav': true},
+                ),
           ],
         ),
         profileTab,
@@ -45,8 +43,20 @@ import '../screens/user-data/routes.dart';
     ),
     userDataRoutes,
     // auth
+
     AutoRoute(page: LoginPage, path: '/login'),
     RedirectRoute(path: '*', redirectTo: '/'),
   ],
 )
 class $RootRouter {}
+
+class InheritedParamScreen extends StatelessWidget {
+  const InheritedParamScreen({Key? key, @pathParam required String id, @queryParam String nonPathParam = 'defa'}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
+
+