+++ auto_route/lib/src/router/widgets/auto_tabs_router.dart
@@ -1,16 +1,14 @@
+import 'package:auto_route/src/router/widgets/auto_tab_view.dart';
 import 'package:collection/collection.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 import '../../../auto_route.dart';
 
-typedef AnimatedIndexedStackBuilder = Widget Function(
-    BuildContext context, Widget child, Animation<double> animation);
-typedef AutoTabsBuilder = Widget Function(
-    BuildContext context, List<Widget> children, TabsRouter tabsRouter);
-typedef AutoTabsPageViewBuilder = Widget Function(
-    BuildContext context, Widget child, PageController pageController);
-typedef AutoTabsTabBarBuilder = Widget Function(
-    BuildContext context, Widget child, TabController tabController);
+typedef AnimatedIndexedStackBuilder = Widget Function(BuildContext context, Widget child, Animation<double> animation);
+typedef AutoTabsBuilder = Widget Function(BuildContext context, List<Widget> children, TabsRouter tabsRouter);
+typedef AutoTabsPageViewBuilder = Widget Function(BuildContext context, Widget child, PageController pageController);
+typedef AutoTabsTabBarBuilder = Widget Function(BuildContext context, Widget child, TabController tabController);
 typedef OnNavigationChanged = Function(TabsRouter tabsRouter);
 
 abstract class AutoTabsRouter extends StatefulWidget {
@@ -28,8 +26,7 @@ abstract class AutoTabsRouter extends StatefulWidget {
     required this.routes,
     this.homeIndex = -1,
     this.inheritNavigatorObservers = true,
-    this.navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    this.navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   }) : super(key: key);
 
   const factory AutoTabsRouter({
@@ -54,7 +51,8 @@ abstract class AutoTabsRouter extends StatefulWidget {
     Curve curve,
     bool inheritNavigatorObservers,
     NavigatorObserversBuilder navigatorObservers,
-    ScrollPhysics? physics,
+     ScrollPhysics? physics,
+     DragStartBehavior dragStartBehavior,
   }) = _AutoTabsRouterPageView;
 
   const factory AutoTabsRouter.tabBar({
@@ -67,6 +65,7 @@ abstract class AutoTabsRouter extends StatefulWidget {
     bool inheritNavigatorObservers,
     NavigatorObserversBuilder navigatorObservers,
     ScrollPhysics? physics,
+    DragStartBehavior dragStartBehavior,
   }) = _AutoTabsRouterTabBar;
 
   const factory AutoTabsRouter.builder({
@@ -84,8 +83,7 @@ abstract class AutoTabsRouter extends StatefulWidget {
     var scope = TabsRouterScope.of(context, watch: watch);
     assert(() {
       if (scope == null) {
-        throw FlutterError(
-            'AutoTabsRouter operation requested with a context that does not include an AutoTabsRouter.\n'
+        throw FlutterError('AutoTabsRouter operation requested with a context that does not include an AutoTabsRouter.\n'
             'The context used to retrieve the AutoTabsRouter must be that of a widget that '
             'is a descendant of an AutoTabsRouter widget.');
       }
@@ -163,8 +161,7 @@ class _AutoTabsRouterIndexedStack extends AutoTabsRouter {
     this.builder,
     int homeIndex = -1,
     bool inheritNavigatorObservers = true,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   }) : super._(
           key: key,
           routes: routes,
@@ -174,19 +171,16 @@ class _AutoTabsRouterIndexedStack extends AutoTabsRouter {
         );
 
   @override
-  _AutoTabsRouterIndexedStackState createState() =>
-      _AutoTabsRouterIndexedStackState();
+  _AutoTabsRouterIndexedStackState createState() => _AutoTabsRouterIndexedStackState();
 }
 
-class _AutoTabsRouterIndexedStackState extends _AutoTabsRouterState
-    with SingleTickerProviderStateMixin {
+class _AutoTabsRouterIndexedStackState extends _AutoTabsRouterState with SingleTickerProviderStateMixin {
   late AnimationController _animationController;
   late Animation<double> _animation;
   int _index = 0;
   late int _tabsHash;
 
-  _AutoTabsRouterIndexedStack get typedWidget =>
-      widget as _AutoTabsRouterIndexedStack;
+  _AutoTabsRouterIndexedStack get typedWidget => widget as _AutoTabsRouterIndexedStack;
 
   @override
   void initState() {
@@ -204,10 +198,10 @@ class _AutoTabsRouterIndexedStackState extends _AutoTabsRouterState
     _tabsHash = const ListEquality().hash(widget.routes);
   }
 
-  @override
+   @override
   void dispose() {
-    _animationController.dispose();
     super.dispose();
+    _animationController.dispose();
   }
 
   @override
@@ -311,14 +305,12 @@ class _IndexedStackBuilder extends StatefulWidget {
   _IndexedStackBuilderState createState() => _IndexedStackBuilderState();
 }
 
-class _IndexedStackBuilderState extends State<_IndexedStackBuilder>
-    with _RouteAwareTabsMixin<_IndexedStackBuilder> {
+class _IndexedStackBuilderState extends State<_IndexedStackBuilder> with _RouteAwareTabsMixin<_IndexedStackBuilder> {
   final _dummyWidget = const SizedBox.shrink();
   final _initializedPagesTracker = <int, bool>{};
 
   @override
-  List<RouteMatch> get routes =>
-      widget.stack.map((e) => e.routeData.route).toList();
+  List<RouteMatch> get routes => widget.stack.map((e) => e.routeData.route).toList();
 
   @override
   List<NavigatorObserver> get observers => widget.navigatorObservers;
@@ -348,8 +340,7 @@ class _IndexedStackBuilderState extends State<_IndexedStackBuilder>
       _setup();
       return;
     }
-    if (widget.lazyLoad &&
-        _initializedPagesTracker[widget.activeIndex] != true) {
+    if (widget.lazyLoad && _initializedPagesTracker[widget.activeIndex] != true) {
       _initializedPagesTracker[widget.activeIndex] = true;
       _didInitTabRoute(widget.activeIndex, oldWidget.activeIndex);
     } else if (widget.activeIndex != oldWidget.activeIndex) {
@@ -366,14 +357,11 @@ class _IndexedStackBuilderState extends State<_IndexedStackBuilder>
       children: List.generate(
         widget.stack.length,
         (index) {
-          if (!widget.stack[index].maintainState &&
-              index != widget.activeIndex) {
+          if (!widget.stack[index].maintainState && index != widget.activeIndex) {
             _initializedPagesTracker[index] = false;
           }
           final isInitialized = _initializedPagesTracker[index] == true;
-          return isInitialized
-              ? widget.itemBuilder(context, index)
-              : _dummyWidget;
+          return isInitialized ? widget.itemBuilder(context, index) : _dummyWidget;
         },
       ),
     );
@@ -385,7 +373,10 @@ class _AutoTabsRouterPageView extends AutoTabsRouter {
   final bool animatePageTransition;
   final Duration duration;
   final Curve curve;
+
+
   final ScrollPhysics? physics;
+  final DragStartBehavior dragStartBehavior;
 
   const _AutoTabsRouterPageView({
     Key? key,
@@ -395,10 +386,10 @@ class _AutoTabsRouterPageView extends AutoTabsRouter {
     this.animatePageTransition = true,
     this.duration = kTabScrollDuration,
     this.curve = Curves.easeInOut,
-    bool inheritNavigatorObservers = true,
     this.physics,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    this.dragStartBehavior = DragStartBehavior.start,
+    bool inheritNavigatorObservers = true,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   })  : _pageViewModeBuilder = builder,
         super._(
           key: key,
@@ -412,48 +403,34 @@ class _AutoTabsRouterPageView extends AutoTabsRouter {
   AutoTabsRouterPageViewState createState() => AutoTabsRouterPageViewState();
 }
 
-class AutoTabsRouterPageViewState extends _AutoTabsRouterState
-    with _RouteAwareTabsMixin<AutoTabsRouter> {
+class AutoTabsRouterPageViewState extends _AutoTabsRouterState with _RouteAwareTabsMixin<AutoTabsRouter> {
   late PageController _pageController;
 
   @override
   void _setupController() {
     assert(_controller != null);
     _controller!.setupRoutes(widget.routes);
-    _pageController = PageController(initialPage: _controller!.activeIndex);
+    _updatePageController();
     _didInitTabRoute(_controller!.activeIndex);
     _controller!.addListener(() {
       if (_controller!.activeIndex != _pageController.page) {
-        _didChangeTabRoute(
-            _controller!.activeIndex, _pageController.page?.toInt() ?? 0);
-      }
-      if (_controller!.activeIndex != _pageController.page?.round()) {
-        if (typedWidget.animatePageTransition &&
-            _canAnimateTransition(
-                _pageController.page, _controller!.activeIndex)) {
-          _pageController.animateToPage(
-            _controller!.activeIndex,
-            duration: typedWidget.duration,
-            curve: typedWidget.curve,
-          );
-        } else {
-          _pageController.jumpToPage(_controller!.activeIndex);
-        }
-      } else if (mounted) {
-        setState(() {});
+        _didChangeTabRoute(_controller!.activeIndex, _controller!.previousIndex!);
       }
     });
+  }
+
+  void _updatePageController() {
+    _pageController = PageController(
+      initialPage: _controller!.activeIndex,
+    );
   }
 
   @override
   void didUpdateWidget(covariant _AutoTabsRouterPageView oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!const ListEquality().equals(widget.routes, oldWidget.routes)) {
-      _controller!.replaceAll(
-          widget.routes, oldWidget.routes[_pageController.page?.round() ?? 0]);
-      _pageController.jumpToPage(
-        _controller!.activeIndex,
-      );
+      _controller!.replaceAll(widget.routes, oldWidget.routes[_controller!.activeIndex]);
+      _updatePageController();
     }
   }
 
@@ -462,7 +439,6 @@ class AutoTabsRouterPageViewState extends _AutoTabsRouterState
   @override
   Widget build(BuildContext context) {
     assert(_controller != null);
-    final stack = _controller!.stack;
     final builder = typedWidget._pageViewModeBuilder ?? _defaultPageViewBuilder;
     final stateHash = controller!.stateHash;
     return RouterScope(
@@ -476,17 +452,11 @@ class AutoTabsRouterPageViewState extends _AutoTabsRouterState
         child: Builder(builder: (context) {
           return builder(
             context,
-            PageView.builder(
-              controller: _pageController,
-              itemCount: stack.length,
+            AutoPageView(
               physics: typedWidget.physics,
-              onPageChanged: _controller!.setActiveIndex,
-              itemBuilder: (context, index) {
-                return KeepAliveTab(
-                  key: ValueKey(index),
-                  page: stack[index],
-                );
-              },
+              dragStartBehavior: typedWidget.dragStartBehavior,
+              controller: _pageController,
+              router: _controller!,
             ),
             _pageController,
           );
@@ -499,19 +469,11 @@ class AutoTabsRouterPageViewState extends _AutoTabsRouterState
     return child;
   }
 
-  // to make sure we don't animate to a page that's more than
-  // one step away
-  bool _canAnimateTransition(double? pageViewIndex, int activeIndex) {
-    if (pageViewIndex == null) return false;
-    return (pageViewIndex - activeIndex).abs() <= 1;
-  }
-
   @override
   List<NavigatorObserver> get observers => _navigatorObservers;
 
   @override
-  List<RouteMatch> get routes =>
-      _controller!.stackData.map((e) => e.route).toList();
+  List<RouteMatch> get routes => _controller!.stackData.map((e) => e.route).toList();
 }
 
 class _AutoTabsRouterTabBar extends AutoTabsRouter {
@@ -519,7 +481,7 @@ class _AutoTabsRouterTabBar extends AutoTabsRouter {
   final Duration? duration;
   final Curve curve;
   final ScrollPhysics? physics;
-
+  final DragStartBehavior dragStartBehavior;
   const _AutoTabsRouterTabBar({
     Key? key,
     required List<PageRouteInfo> routes,
@@ -528,9 +490,9 @@ class _AutoTabsRouterTabBar extends AutoTabsRouter {
     this.duration,
     this.curve = Curves.ease,
     bool inheritNavigatorObservers = true,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
     this.physics,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    this.dragStartBehavior = DragStartBehavior.start,
   }) : super._(
           key: key,
           routes: routes,
@@ -543,14 +505,34 @@ class _AutoTabsRouterTabBar extends AutoTabsRouter {
   _AutoTabsRouterTabBarState createState() => _AutoTabsRouterTabBarState();
 }
 
-class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
-    with _RouteAwareTabsMixin<AutoTabsRouter>, TickerProviderStateMixin {
+class _AutoTabsRouterTabBarState extends _AutoTabsRouterState with _RouteAwareTabsMixin<AutoTabsRouter>, TickerProviderStateMixin {
   late TabController _tabController;
 
   @override
   void _setupController() {
     assert(_controller != null);
     _controller!.setupRoutes(widget.routes);
+    _updateTabController();
+    _didInitTabRoute(_controller!.activeIndex);
+    _controller!.addListener(() {
+      if (_controller!.activeIndex != _tabController.index) {
+        _didChangeTabRoute(
+          _controller!.activeIndex,
+          _tabController.index,
+        );
+        _tabController.animateTo(
+          _controller!.activeIndex,
+          duration: typedWidget.duration,
+          curve: typedWidget.curve,
+        );
+      }
+      if (mounted) {
+        setState(() {});
+      }
+    });
+  }
+
+  void _updateTabController() {
     _tabController = TabController(
       initialIndex: _controller!.activeIndex,
       length: _controller!.pageCount,
@@ -559,36 +541,14 @@ class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
     _tabController.addListener(() {
       _controller!.setActiveIndex(_tabController.index);
     });
-    _didInitTabRoute(_controller!.activeIndex);
-    _controller!.addListener(() {
-      if (_controller!.activeIndex != _controller!.previousIndex) {
-        _didChangeTabRoute(_controller!.activeIndex,
-            _controller!.previousIndex ?? _tabController.index);
-      }
-      if (_controller!.activeIndex != _tabController.index &&
-          !_tabController.indexIsChanging) {
-        _tabController.animateTo(
-          _controller!.activeIndex,
-          duration: typedWidget.duration,
-          curve: typedWidget.curve,
-        );
-      } else if (mounted) {
-        setState(() {});
-      }
-    });
   }
 
   @override
   void didUpdateWidget(covariant _AutoTabsRouterTabBar oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!const ListEquality().equals(widget.routes, oldWidget.routes)) {
-      _controller!
-          .replaceAll(widget.routes, oldWidget.routes[_tabController.index]);
-      _tabController.animateTo(
-        _controller!.activeIndex,
-        curve: typedWidget.curve,
-        duration: typedWidget.duration,
-      );
+      _controller!.replaceAll(widget.routes, oldWidget.routes[_tabController.index]);
+      _updateTabController();
     }
   }
 
@@ -597,7 +557,6 @@ class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
   @override
   Widget build(BuildContext context) {
     assert(_controller != null);
-    final stack = _controller!.stack;
     final builder = typedWidget.builder ?? _defaultPageViewBuilder;
     final stateHash = controller!.stateHash;
     return RouterScope(
@@ -611,16 +570,11 @@ class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
         child: Builder(builder: (context) {
           return builder(
             context,
-            CustomTabBarView(
-              controller: _tabController,
+            AutoTabView(
               physics: typedWidget.physics,
-              children: List.generate(
-                stack.length,
-                (index) => KeepAliveTab(
-                  key: ValueKey(index),
-                  page: stack[index],
-                ),
-              ),
+              dragStartBehavior: typedWidget.dragStartBehavior,
+              controller: _tabController,
+              router: _controller!,
             ),
             _tabController,
           );
@@ -637,8 +591,7 @@ class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
   List<NavigatorObserver> get observers => _navigatorObservers;
 
   @override
-  List<RouteMatch> get routes =>
-      _controller!.stackData.map((e) => e.route).toList();
+  List<RouteMatch> get routes => _controller!.stackData.map((e) => e.route).toList();
 }
 
 class _AutoTabsRouterBuilder extends AutoTabsRouter {
@@ -654,8 +607,7 @@ class _AutoTabsRouterBuilder extends AutoTabsRouter {
     required this.builder,
     int homeIndex = -1,
     bool inheritNavigatorObservers = true,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   }) : super._(
           key: key,
           routes: routes,
@@ -668,8 +620,7 @@ class _AutoTabsRouterBuilder extends AutoTabsRouter {
   _AutoTabsRouterBuilderState createState() => _AutoTabsRouterBuilderState();
 }
 
-class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
-    with _RouteAwareTabsMixin<AutoTabsRouter> {
+class _AutoTabsRouterBuilderState extends _AutoTabsRouterState with _RouteAwareTabsMixin<AutoTabsRouter> {
   @override
   void _setupController() {
     assert(_controller != null);
@@ -678,8 +629,7 @@ class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
     _didInitTabRoute(_controller!.activeIndex);
     _controller!.addListener(() {
       if (_controller!.activeIndex != _controller!.previousIndex) {
-        _didChangeTabRoute(
-            _controller!.activeIndex, _controller!.previousIndex ?? 0);
+        _didChangeTabRoute(_controller!.activeIndex, _controller!.previousIndex ?? 0);
         typedWidget.onNavigate?.call(_controller!);
       }
       if (mounted) {
@@ -692,8 +642,7 @@ class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
   void didUpdateWidget(covariant _AutoTabsRouterBuilder oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!const ListEquality().equals(widget.routes, oldWidget.routes)) {
-      _controller!.replaceAll(
-          widget.routes, oldWidget.routes[_controller!.activeIndex]);
+      _controller!.replaceAll(widget.routes, oldWidget.routes[_controller!.activeIndex]);
       typedWidget.onNavigate?.call(_controller!);
     }
   }
@@ -735,8 +684,7 @@ class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
   List<NavigatorObserver> get observers => _navigatorObservers;
 
   @override
-  List<RouteMatch> get routes =>
-      _controller!.stackData.map((e) => e.route).toList();
+  List<RouteMatch> get routes => _controller!.stackData.map((e) => e.route).toList();
 }
 
 mixin _RouteAwareTabsMixin<T extends StatefulWidget> on State<T> {
@@ -748,8 +696,7 @@ mixin _RouteAwareTabsMixin<T extends StatefulWidget> on State<T> {
     observers.whereType<AutoRouterObserver>().forEach((observer) {
       TabPageRoute? previousRoute;
       if (previous != -1) {
-        previousRoute =
-            TabPageRoute(routeInfo: routes[previous], index: previous);
+        previousRoute = TabPageRoute(routeInfo: routes[previous], index: previous);
       }
       observer.didInitTabRoute(
         TabPageRoute(routeInfo: routes[index], index: index),
@@ -779,8 +726,7 @@ class KeepAliveTab extends StatefulWidget {
   State<KeepAliveTab> createState() => _KeepAliveTabState();
 }
 
-class _KeepAliveTabState extends State<KeepAliveTab>
-    with AutomaticKeepAliveClientMixin {
+class _KeepAliveTabState extends State<KeepAliveTab> with AutomaticKeepAliveClientMixin {
   @override
   Widget build(BuildContext context) {
     super.build(context);