+++ auto_route_generator/lib/src/models/route_config.dart
@@ -202,7 +202,7 @@ class RouteConfig {
   }
 
   List<ParamConfig> get pathQueryParams {
-    return parameters.where((p) => p.isPathParam || p.isQueryParam).toList();
+    return parameters.where((p) => (p.isPathParam || p.isQueryParam) && !p.isInheritedPathParam).toList();
   }
 
   Iterable<ParamConfig> get requiredParams =>
@@ -242,6 +242,10 @@ class RouteConfig {
 
   bool get hasUnparsableRequiredArgs => parameters.any((p) =>
       (p.isRequired || p.isPositional) && !p.isPathParam && !p.isQueryParam);
+
+
+
+
 }
 
 class MetaEntry<T> {