+++ auto_route_generator/lib/src/models/router_config.dart
@@ -3,7 +3,7 @@ import 'route_config.dart';
 
 class RouterConfig {
   final List<RouteConfig> routes;
-  final RouteConfig globalRouteConfig;
+  final RouteConfig parentRouteConfig;
   final String routerClassName;
   final RouterConfig? parent;
   final String? replaceInRouteName;
@@ -12,7 +12,7 @@ class RouterConfig {
   RouterConfig({
     required this.routes,
     required this.element,
-    required this.globalRouteConfig,
+    required this.parentRouteConfig,
     required this.routerClassName,
     this.parent,
     this.replaceInRouteName,
@@ -21,7 +21,7 @@ class RouterConfig {
   RouterConfig copyWith({
     bool? generateNavigationHelper,
     List<RouteConfig>? routes,
-    RouteConfig? globalRouteConfig,
+    RouteConfig? parentRouteConfig,
     String? routesClassName,
     String? routeNamePrefix,
     String? routerClassName,
@@ -31,7 +31,7 @@ class RouterConfig {
   }) {
     return RouterConfig(
       routes: routes ?? this.routes,
-      globalRouteConfig: globalRouteConfig ?? this.globalRouteConfig,
+      parentRouteConfig: parentRouteConfig ?? this.parentRouteConfig,
       routerClassName: routerClassName ?? this.routerClassName,
       replaceInRouteName: replaceInRouteName ?? this.replaceInRouteName,
       parent: parent ?? this.parent,